+++ component/loki/relabel/relabel_test.go
@@ -10,8 +10,10 @@ import (
 	"github.com/grafana/agent/component"
 	"github.com/grafana/agent/component/common/loki"
 	flow_relabel "github.com/grafana/agent/component/common/relabel"
+	"github.com/grafana/agent/pkg/flow/componenttest"
 	"github.com/grafana/agent/pkg/flow/logging"
 	"github.com/grafana/agent/pkg/river"
+	"github.com/grafana/agent/pkg/util"
 	"github.com/grafana/loki/pkg/logproto"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/common/model"
@@ -276,6 +278,52 @@ func TestCache(t *testing.T) {
 		require.True(t, ok)
 		require.Equal(t, f, wantKeys[i])
 	}
+}
+
+func TestRuleGetter(t *testing.T) {
+	// Set up the component Arguments.
+	originalCfg := `rule {
+         action       = "keep"
+		 source_labels = ["__name__"]
+         regex        = "up"
+       }
+		forward_to = []`
+	var args Arguments
+	require.NoError(t, river.Unmarshal([]byte(originalCfg), &args))
+
+	// Set up and start the component.
+	tc, err := componenttest.NewControllerFromID(util.TestLogger(t), "loki.relabel")
+	require.NoError(t, err)
+	go func() {
+		err = tc.Run(componenttest.TestContext(t), args)
+		require.NoError(t, err)
+	}()
+	require.NoError(t, tc.WaitExports(time.Second))
+
+	// Use the getter to retrieve the original relabeling rules.
+	exports := tc.Exports().(Exports)
+	gotOriginal := exports.Rules()
+
+	// Update the component with new relabeling rules and retrieve them.
+	updatedCfg := `rule {
+         action       = "drop"
+		 source_labels = ["__name__"]
+         regex        = "up"
+       }
+		forward_to = []`
+	require.NoError(t, river.Unmarshal([]byte(updatedCfg), &args))
+
+	require.NoError(t, tc.Update(args))
+	gotUpdated := exports.Rules()
+
+	require.NotEqual(t, gotOriginal, gotUpdated)
+	require.Len(t, gotOriginal, 1)
+	require.Len(t, gotUpdated, 1)
+
+	require.Equal(t, gotOriginal[0].Action, relabel.Keep)
+	require.Equal(t, gotUpdated[0].Action, relabel.Drop)
+	require.Equal(t, gotUpdated[0].SourceLabels, gotOriginal[0].SourceLabels)
+	require.Equal(t, gotUpdated[0].Regex, gotOriginal[0].Regex)
 }
 
 func getEntry() loki.Entry {