+++ component/prometheus/relabel/relabel.go
@@ -42,6 +42,7 @@ type Arguments struct {
 // Exports holds values which are exported by the prometheus.relabel component.
 type Exports struct {
 	Receiver storage.Appendable `river:"receiver,attr"`
+	Rules    flow_relabel.Rules `river:"rules,attr"`
 }
 
 // Component implements the prometheus.relabel component.
@@ -129,7 +130,7 @@ func New(o component.Options, args Arguments) (*Component, error) {
 
 	// Immediately export the receiver which remains the same for the component
 	// lifetime.
-	o.OnStateChange(Exports{Receiver: c.receiver})
+	o.OnStateChange(Exports{Rules: c.getRules})
 
 	// Call to Update() to set the relabelling rules once at the start.
 	if err = c.Update(args); err != nil {
@@ -154,7 +155,6 @@ func (c *Component) Update(args component.Arguments) error {
 	c.clearCache()
 	c.mrc = flow_relabel.ComponentToPromRelabelConfigs(newArgs.MetricRelabelConfigs)
 	c.fanout.UpdateChildren(newArgs.ForwardTo)
-	c.opts.OnStateChange(Exports{Receiver: c.receiver})
 
 	return nil
 }
@@ -234,4 +234,11 @@ func (c *Component) addToCache(originalID uint64, lbls labels.Labels) {
 type labelAndID struct {
 	labels labels.Labels
 	id     uint64
+}
+
+func (c *Component) getRules() []*relabel.Config {
+	c.mut.RLock()
+	defer c.mut.RUnlock()
+
+	return c.mrc
 }