+++ CHANGELOG.md
@@ -34,8 +34,12 @@ Main (unreleased)
   - `otelcol.receiver.kafka` receives telemetry data from Kafka. (@rfratto)
   - `phlare.scrape` collects application performance profiles. (@cyriltovena)
   - `phlare.write` sends application performance profiles to Grafana Phlare. (@cyriltovena)
-
   - `otelcol.receiver.zipkin` receives Zipkin-formatted traces. (@rfratto)
+
+- Flow components which work with relabeling rules (`discovery.relabel`,
+  `prometheus.relabel` and `loki.relabel`) now export a new field named Rules.
+  The exported value is a method that returns the currently configured rules.
+  (@tpaschalis)
 
 ### Enhancements